added cart routes

+++ api/dao/CartDao.class.php
@@ -9,35 +9,51 @@ class CartDao extends BaseDao
 
   public function get_medicine_in_cart_by_account($id){
     return $this->query_unique("SELECT * FROM carts
-                          WHERE status != :status AND account_id = :id", ["id" => $id, "status" => "BOUGHT"]);
-  }
-
-  public function get_medicine_in_carts(){
-    return $this->get_all();
+                          WHERE status != :status AND account_id = :id", ["id" => $id, "status" => "BOUGHT"]); //show only one that is not bought!!!
   }
 
   public function get_total_price_by_account($id){
-    return $this->query_unique("SELECT SUM(m.price) FROM carts c, medicines m
+     $data = $this->query_unique("SELECT SUM(m.price * c.quantity) AS total FROM carts c, medicines m
                           WHERE c.medicine_id= m.id AND account_id = :id", ["id" => $id]);
+     if($data['total'] != null)
+        return $data['total'];
+    else return "0.00";
   }
 
-  public function alter_cart_by_account($account, $medicine){ //will not use this
-    $query = "DELETE FROM carts WHERE account_id= :account_id AND medicine_id= :medicine_id";
-    $stmt = $this->connection->prepare($query);
-    $params=["account_id" => $account, "medicine_id" => $medicine];
-    $stmt -> execute($params);
+   public function alter_cart_by_account($account, $medicine){
+    $data = $this->get_medicine($account, $medicine);
+
+    if($data != null){                                                                                          //user can remove his itmes
+        $query = "DELETE FROM carts WHERE account_id= :account_id AND medicine_id= :medicine_id";
+        $stmt = $this->connection->prepare($query);
+        $params=["account_id" => $account, "medicine_id" => $medicine];
+        $stmt -> execute($params);
+    } else {
+      throw new \Exception("Not found in cart", 404);
+    }
   }
 
-  public function update_status($account, $medicine, $status){ //change status insted of delete
-    $query = "UPDATE carts SET status = :status WHERE account_id= :account_id AND medicine_id= :medicine_id";
-    $stmt = $this->connection->prepare($query);
-    $params=["status" => $status, "account_id" => $account, "medicine_id" => $medicine];
-    $stmt -> execute($params);
-  }
+ public function update_status($account, $status){                                                               //change status to bought
+   $data = $this->query_unique("SELECT * FROM carts
+                         WHERE account_id = :id", ["id" => $account]);
+    if($data != null){
+      $query = "UPDATE carts SET status = :status WHERE account_id= :account_id";
+      $stmt = $this->connection->prepare($query);
+      $params=["status" => $status, "account_id" => $account];
+      $stmt -> execute($params);
+    } else {
+      throw new \Exception("Nothing in cart", 404);
+
+    }
 
   }
 
 
-
+private function get_medicine($account, $medicine){
+  $data = $this->query_unique("SELECT * FROM carts
+                        WHERE medicine_id = :medicine AND account_id = :id", ["id" => $account, "medicine" => $medicine]);
+  return $data;
+ }
+}
 
 ?>
